move barcode files into barcode module to resolve reference issues
